requirements.txt
Django>=4.2.11
djangorestframework>=3.15.1
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.1
python-decouple>=3.8
dj-database-url>=2.1.0
psycopg2-binary>=2.9.10
supabase>=2.4.0
Pillow>=11.0.0
twilio>=8.13.0
google-auth>=2.29.0
requests>=2.31.0
gunicorn>=21.2.0
whitenoise>=6.6.0
